+++ app/routes/content.js
@@ -1,5 +1,6 @@
 import { isArray } from '@ember/array';
 import Route from '@ember/routing/route';
+import DS from 'ember-data';
 
 // Error handling for API
 const handlers = new Map([
@@ -27,11 +28,15 @@ export default Route.extend({
     actions: {
         error(error) {
             // Handle API Errors
-            if (error && error.errors && isArray(error.errors)) {
+            if (error && !(error instanceof DS.AbortError)
+                && error.errors && isArray(error.errors)) {
+                // If  the error is not a AbortError (no connection), we handle it here.
                 const { detail } = error.errors[0];
                 const page = handlers.get(detail) || 'page-not-found';
-
                 return this.intermediateTransitionTo(page);
+            } else {
+                // Otherwise, we bubble it to the error handler in our parent route.
+                return true;
             }
         },
     },