+++ app/routes/content/index.js
@@ -91,7 +91,6 @@ export default Route.extend(Analytics, ResetScrollMixin, {
             if (error && error.errors && isArray(error.errors)) {
                 const { detail } = error.errors[0];
                 const page = handlers.get(detail) || 'page-not-found';
-
                 return this.intermediateTransitionTo(page);
             }
         },
@@ -170,7 +169,6 @@ export default Route.extend(Analytics, ResetScrollMixin, {
         const facebookAppId = provider.get('facebookAppId') || config.FB_APP_ID;
         const mintDoi = extractDoiFromString(preprint.get('preprintDoiUrl'));
         const peerDoi = preprint.get('doi');
-        const doi = peerDoi || mintDoi;
         const image = this.get('theme.logoSharing');
         const imageUrl = /^https?:\/\//.test(image.path) ? image.path : origin + image.path;
         const dateCreated = new Date(preprint.get('dateCreated') || null);
@@ -207,8 +205,8 @@ export default Route.extend(Analytics, ResetScrollMixin, {
             ['citation_public_url', canonicalUrl],
             ['citation_online_date', `${dateCreated.getFullYear()}/${dateCreated.getMonth() + 1}/${dateCreated.getDate()}`],
         ];
-        if (doi) {
-            highwirePress.push(['citation_doi', doi]);
+        if (mintDoi) {
+            highwirePress.push(['citation_doi', mintDoi]);
         }
 
         // TODO map Eprints fields